### 2019-05-20 #1464 @danielskatz
👋 @martibosch - you need to add single quotes `'` around the title in your paper.md

Once you have done so, you can enter ```@whedon generate pdf``` as a new comment here to see if that fixes the build problem
