### 2019-05-24 #1464 @martibosch
Thank you very much @leouieda

A small update: a preprint which shows how the library might be used in real studies is already available (https://doi.org/10.1101/645549), and so is the repository with the materials required to reproduce the results (https://github.com/martibosch/swiss-urbanization). Should we add this to the manuscript?

On the other hand, the library has also been used in a manuscript submitted to the "Landscape Ecology"  (Springer) journal, which is currently undergoing major modifications. Although we have not published any preprint, the code to reproduce our results (using this library) is also available at a dedicated repository (https://github.com/martibosch/carbon-sequestration-vaud).

Martí
