### 2019-06-05 #1464 @leouieda
@martibosch you could add a link to these repositories as examples of the software package. But they don't need to be on the paper (maybe just as a citation to the preprint). The JOSS papers are usually very minimal and most of your efforts should go into the software documentation instead.
