### 2019-06-18 #1511 @whedon
[REVIEW]: swisslandstats-geopy: Python tools for the land statistics datasets from the Swiss Federal Statistical Office
**Submitting author:** @martibosch (<a href="http://orcid.org/0000-0001-8735-9144">Martí Bosch</a>)
**Repository:** <a href="https://github.com/martibosch/swisslandstats-geopy" target ="_blank">https://github.com/martibosch/swisslandstats-geopy</a>
**Version:** 0.7.2
