### 2019-06-19 #1511 @leouieda
:wave: @weikang9009 @darribas @martibosch this is the review thread for the `swisslandstats-geopy` paper. All of our communications will happen here from now on.

Both @weikang9009 and @darribas have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

We aim for reviews to be completed within 2-4 weeks. Please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me here (@leouieda) or email me privately if you have any questions/concerns.
