### 2019-07-05 martibosch/swisslandstats-geopy#2 @weikang9009
Latest github release v0.1 does not match the latest pypi release v0.6.1 
[v0.6.1 is tagged but not released on github](https://github.com/martibosch/swisslandstats-geopy/releases)

openjournals/joss-reviews#1511
