### 2019-07-06 martibosch/swisslandstats-geopy#3 @weikang9009
Missing API method documentation
I would suggest some additional work on API documentation for the general public (sth like [the API references page of seaborn](https://seaborn.pydata.org/api.html)). A documentation website based on [readthedocs](https://readthedocs.org/) would be great.

This issue is related to the review of the JOSS submission https://github.com/openjournals/joss-reviews/issues/1511
`Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?`
