### 2019-07-06 76daa7b
added docs + readthedocs config + readme badge
