+++ docs/src/swisslandstats.rst
@@ -1,0 +1,15 @@
+swisslandstats package
+======================
+
+LandDataFrame class
+-------------------
+
+.. autoclass:: swisslandstats.LandDataFrame
+   :members:
+
+Other functions
+---------------
+
+.. autofunction:: swisslandstats.merge
+      
+.. autofunction:: swisslandstats.read_csv
