### 2019-07-06 martibosch/swisslandstats-geopy#3 @martibosch
Hello,
the main functionalities are documented as docstrings of their respective methods. Nonetheless, I have compiled them via Sphinx into a [dedicated readthedocs website](https://swisslandstats-geopy.readthedocs.io/). I hope that this addresses the issue.

Martí
