### 2019-07-06 martibosch/swisslandstats-geopy#2 @martibosch
Hello,

I have manually added a release for the latest tag (v0.6.1), and I will do so for the following releases.

Martí
