### 2019-07-07 martibosch/swisslandstats-geopy#2 @weikang9009
Thanks! Looks good.
