### 2019-07-08 martibosch/swisslandstats-geopy#3 @weikang9009
Hi Martí, the documentation website looks nice, but it seems you only expose `LandDataFrame` class on the website https://swisslandstats-geopy.readthedocs.io/en/latest/swisslandstats.html while other functions (like [`read_csv` function](https://github.com/martibosch/swisslandstats-geopy/blob/master/swisslandstats/dataframe.py)) are not listed on the docs website?
