### 2019-07-08 martibosch/swisslandstats-geopy#4 @weikang9009
Plot function/method does not give the correct `x` and `y` coordinates?
The `plot` method for the `LandDataFrame` class  does not visualize the correct `x` and `y` coordinates as shown below (this is a screenshot from the "overview" notebook):
![image](https://user-images.githubusercontent.com/7359284/60834715-27f39b00-a176-11e9-9d37-c6d8a5c97bc4.png)
The plotted figure should have the same`x` and `y` coordinates as the following figure which is produced from `rasterio`'s plotting function:
rasterio
![image](https://user-images.githubusercontent.com/7359284/60835231-73f30f80-a177-11e9-95aa-2a023ce53144.png)

this issue is related to https://github.com/openjournals/joss-reviews/issues/1511
