### 2019-07-09 martibosch/swisslandstats-geopy#3 @martibosch
Hello Wei,
I have amended the latest commit and included the documentation for the `merge` and `read_csv` functions. I have ommited creating documentation pages for the `plotting` and `geometry` modules on purpose, since all their respective functions are embedded as methods of the `LandDataFrame` class, and therefore their documentation can already be found there. I believe that such modularization of the library is appropriate for the maintainability of the code but does not concern so much the end user (the main target for whom the documentation is provided).
