+++ setup.py
@@ -36,7 +36,7 @@ def setup_package():
     # yapf: disable
     setup(
         name='swisslandstats-geopy',
-        version='0.1.0',
+        version='0.2.0',
         description='Python for the land statistics datasets from the SFSO',
         long_description=long_description,
         long_description_content_type='text/markdown',
