+++ setup.cfg
@@ -1,11 +1,11 @@
+[bumpversion]
+current_version = 0.2.0
+commit = True
+tag = True
+
 [yapf]
 based_on_style = pep8
 split_before_named_assigns = false
-
-[bumpversion]
-current_version = 0.1.0
-commit = True
-tag = True
 
 [bumpversion:file:setup.py]
 search = version='{current_version}'
@@ -14,3 +14,4 @@ replace = version='{new_version}'
 [bumpversion:file:swisslandstats/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
+
