### 2019-07-09 6a04890
plotting with an affine transform via rasterio.plot.show
