+++ setup.py
@@ -40,7 +40,7 @@ def setup_package():
     # yapf: disable
     setup(
         name='swisslandstats-geopy',
-        version='0.6.1',
+        version='0.7.0',
         description='Python for the land statistics datasets from the SFSO',
         long_description=long_description,
         long_description_content_type='text/markdown',
