### 2019-07-09 09bdb36
fix: deleted reqs-dev from setup.py (not included in MANIFEST)
