### 2019-07-09 martibosch/swisslandstats-geopy#4 @martibosch
Hello Wei,
I have changed the code of `swisslandstats.plotting.plot_ndarray` (in turn used in `LandDataFrame.plot`) so that it uses `rasterio.plot.show` (instead of `matplotlib.pyplot.imshow`). This is released in v0.7.0 (I have also created [a change log](https://github.com/martibosch/swisslandstats-geopy/blob/master/CHANGELOG.md)). I have also updated the [example notebook](https://github.com/martibosch/swisslandstats-notebooks/blob/master/overview.ipynb) with the transformed plot coordinates.

Martí
