### 2019-07-09 martibosch/swisslandstats-geopy#3 @weikang9009
I agree with you. The docs website looks great now!
