### 2019-07-09 martibosch/swisslandstats-geopy#5 @weikang9009
Community guidelines
Please add community guidelines: clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

related to https://github.com/openjournals/joss-reviews/issues/1511
