### 2019-07-09 martibosch/swisslandstats-geopy#6 @weikang9009
update figures in README.md
[README.md](https://github.com/martibosch/swisslandstats-geopy/blob/master/README.md) figures need to be updated  (due to the change in the plotting methods https://github.com/martibosch/swisslandstats-geopy/issues/4)

related to https://github.com/openjournals/joss-reviews/issues/1511
