### 2019-07-10 4a407fa
updated README example plots (axis coords with rasterio.plot.show)
