### 2019-07-10 martibosch/swisslandstats-geopy#6 @martibosch
Done in  [cd5fb12](https://github.com/martibosch/swisslandstats-geopy/commit/cd5fb12c103e91d57d4c09d3e301270c6abf0aa3) :heavy_check_mark:
