### 2019-07-10 acc24d5
[ci skip] added contributing guide and github issue template
