+++ CONTRIBUTING.md
@@ -1,0 +1,35 @@
+# Contributing
+
+Contributions are always greatly appreciated and credit will always be given.
+
+## Types of contributions
+
+### Report bugs
+
+If you've found an error:
+
+* Read the error message and documentation
+* Search through the open issues and closed issues first
+* If the problem is with a dependency of this project, open an issue on the dependency's repo
+* If the problem is with swisslandstats and you can fix it simply, please submit a pull request
+* If the problem persists, please report the bug at https://github.com/martibosch/swisslandstats-notebooks/issues and include:
+
+    * Your operating system name and version.
+    * Any details about your local setup that might be helpful in troubleshooting.
+    * Detailed steps to reproduce the bug.
+
+### Fix bugs
+
+Look through the GitHub issues for bugs. Anything tagged with "bug" and "help wanted" is open to whoever wants to implement it.
+
+### Implement features
+
+Look through the GitHub issues for features. Anything tagged with "enhancement" and "help wanted" is open to whoever wants to implement it.
+
+## Pull request guidelines
+
+Before you submit a pull request, check that it meets these guidelines:
+
+1. The pull request should include tests.
+2. If the pull request adds functionality, the docs should be updated. Put your new functionality into a function with a docstring, and add the feature to the list in README.md.
+3. The pull request should work for Python 2.7 and 3.6, as well as for PyPy. Check https://travis-ci.org/martibosch/pylandstats/pull_requests and make sure that the tests pass for all supported Python versions.
