+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,20 @@
+This issue tracker is for reporting bugs or proposing new features. Please ask "how-to" questions on https://stackoverflow.com/ instead. Before you proceed, please read the contributing guidelines in this repo's CONTRIBUTING.md.
+
+If you're having trouble with swisslandstats, first search the previously opened issues (and Stack Overflow) to see if the problem has already been noted. If not, fill in the template below:
+
+#### Problem description (what did you do, what did you expect to happen, and what actually happened)
+
+#### What operating system, architecture, Python version, and OSMnx version are you using?
+
+#### Complete list of your environment's packages and their versions (for example, run `conda list` or `pip list` then paste the output below, between the two "details" tags)
+
+<details>
+# Paste the output of your Python packages and their versions here, between these two "details" tags
+
+</details>
+
+#### Code snippet that independently and completely reproduces the issue
+
+```python
+# your code goes here
+```
