### 2019-07-10 martibosch/swisslandstats-geopy#5 @martibosch
A [contribution guide](https://github.com/martibosch/swisslandstats-geopy/blob/master/CONTRIBUTING.md) and a [github issue template](https://github.com/martibosch/swisslandstats-geopy/blob/master/.github/ISSUE_TEMPLATE.md) have been added in [acc24d5](https://github.com/martibosch/swisslandstats-geopy/commit/acc24d5e596582913ca40336fec38a2479d2c6e4)
