### 2019-07-10 #1511 @leouieda
:wave: Hi @weikang9009 and @darribas just checking in on the progress of the reviews. I see some issues have been opened (and resolved) in the project repository. Please let me know what your current status is and any hangups in the review that might need to be resolved.
