### 2019-07-11 martibosch/swisslandstats-geopy#7 @weikang9009
target audience
Could you explain who the target audience is in the documentation as well as in the paper? I

https://github.com/openjournals/joss-reviews/issues/1511
