### 2019-07-11 martibosch/swisslandstats-geopy#8 @weikang9009
Reference DOI
try to update the references in the paper to fulfill the requirement:

References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

https://github.com/openjournals/joss-reviews/issues/1511
