### 2019-07-11 #1511 @weikang9009
Hi @leouieda, I am almost done with the review. @martibosch has been very responsive in resolving the issues I opened. I just opened two issues related to adding a statement of the target audience and missing DOIs in the paper.
