### 2019-07-11 #1511 @leouieda
@weikang9009 thanks for the update and opening all those issues :+1: Let me know when you have a final verdict (meaning that the software passes all requirements)
