### 2019-07-14 6ac4b28
[ci skip] updated paper references with DOIs
