+++ paper/paper.bib
@@ -5,14 +5,6 @@
   howpublished={Available from \url{https://www.python.org/dev/peps/pep-0257/}. Accessed 20 May 2019},
 }
 
-@book{oliphant2006guide,
-  title={A guide to NumPy},
-  author={Oliphant, Travis E},
-  volume={1},
-  year={2006},
-  publisher={Trelgol Publishing USA}
-}
-
 @misc{gillies2007shapely,
   title={Shapely: manipulation and analysis of geometric objects},
   author={Sean Gillies and others},
@@ -30,6 +22,18 @@
   organization={Austin, TX}
 }
 
+@article{van2011numpy,
+  title={The NumPy array: a structure for efficient numerical computation},
+  author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
+  journal={Computing in Science \& Engineering},
+  volume={13},
+  number={2},
+  pages={22},
+  year={2011},
+  publisher={IEEE Computer Society},
+  doi={10.1109/MCSE.2011.37}
+}
+
 @article{boeing2017osmnx,
   title={OSMnx: New methods for acquiring, constructing, analyzing, and visualizing complex street networks},
   author={Boeing, Geoff},
@@ -37,7 +41,8 @@
   volume={65},
   pages={126--139},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compenvurbsys.2017.05.004}
 }
 
 @misc{sfso2017statistique,
@@ -53,4 +58,5 @@
   booktitle={Proceedings of the 17th Python in Science Conference},
   pages={113-120},
   year={2018},
+  doi={10.25080/Majora-4af1f417-011}
 }
