### 2019-07-14 fabbb2e
[ci skip] added RTD site, removed docstrings reference
