+++ paper/paper.bib
@@ -1,10 +1,3 @@
-@misc{goodger2001docstring,
-  title={Docstring Conventions},
-  author={Goodger, David and van Rossum, Guido},
-  year={2001},
-  howpublished={Available from \url{https://www.python.org/dev/peps/pep-0257/}. Accessed 20 May 2019},
-}
-
 @misc{gillies2007shapely,
   title={Shapely: manipulation and analysis of geometric objects},
   author={Sean Gillies and others},
