### 2019-07-14 e560582
[ci skip] added target audience and refs to literature applications
