+++ paper/paper.bib
@@ -53,3 +53,25 @@
   year={2018},
   doi={10.25080/Majora-4af1f417-011}
 }
+
+@article{jaligot2019assessing,
+  title={Assessing spatial temporal patterns of ecosystem services in Switzerland},
+  author={Jaligot, R{\'e}mi and Chenal, J{\'e}r{\^o}me and Bosch, Mart{\'\i}},
+  journal={Landscape Ecology},
+  volume={34},
+  number={6},
+  pages={1379--1394},
+  year={2019},  
+  publisher={Springer},
+  doi={10.1007/s10980-019-00850-7}
+}
+
+@article{bosch2019spatiotemporal,
+  title={Spatiotemporal patterns of urbanization in three Swiss urban agglomerations: insights from landscape metrics, growth modes and fractal analysis},
+  author={Bosch, Mart{\'\i} and Chenal, J{\'e}r{\^o}me},
+  journal={BioRxiv},
+  pages={645549},
+  year={2019},
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/645549}
+}
