+++ README.md
@@ -7,7 +7,9 @@
 
 # swisslandstats-geopy
 
-Extended pandas-like interface for the [Swiss Land Statistics datasets from the Swiss Federal Statistical Office](https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/sol-utilisation-couverture/statistique-suisse-superficie.html) (link in French)
+Extended pandas-like interface for the [Swiss Land Statistics datasets from the Swiss Federal Statistical Office (SFSO)](https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/sol-utilisation-couverture/statistique-suisse-superficie.html) (link in French)
+
+The target audience of `swisslandstats-geopy` is researchers and developers in environmental sciences and GIS, who intend to produce repeatable and reproducible computational workflows that make use of the land use/land cover inventory provided by the SFSO. 
 
 ## Features
 
@@ -30,7 +32,11 @@ vaud_ldf.plot('AS09R_4', cmap=sls.noas04_4_cmap, legend=True)
 ```
 ![landstats-vaud](examples/landstats_vaud.png)
 
-See the [swisslandstats-notebooks](https://github.com/martibosch/swisslandstats-notebooks) repository for a more thorough overview.
+See the [swisslandstats-notebooks](https://github.com/martibosch/swisslandstats-notebooks) repository for a more thorough overview. Examples of applications of the library in the academic literature include:
+
+* The assessment of the carbon sequestration for the canton of Vaud (see [the dedicated GitHub repository](https://github.com/martibosch/carbon-sequestration-vaud) with the materials necessary to reproduce the results)
+* The evaluation of the spatio-temporal patterns of LULC change in the urban agglomerations of Zurich, Bern and Lausanne (see [the dedicated GitHub repository](https://github.com/martibosch/swiss-urbanization) with the materials necessary to reproduce the results).
+
 
 ## Installation
 
