### 2019-07-14 f72088b
[ci skip] deleted affiliation index since there is only one
