+++ paper/paper.md
@@ -9,13 +9,10 @@ tags:
 authors:
  - name: Martí Bosch
    orcid: 0000-0001-8735-9144
-   affiliation: 1
  - name: Jérôme Chenal
    orcid: 0000-0002-8109-8358
-   affiliation: 1
 affiliations:
  - name: Urban and Regional Planning Community (CEAT), École Polytechnique Fédérale de Lausanne (EPFL), Switzerland
-   index: 1
 date: 17 May 2019
 bibliography: paper.bib
 ---
