### 2019-07-14 martibosch/swisslandstats-geopy#8 @martibosch
Hello,
I hope that this is addressed in [6ac4b28](https://github.com/martibosch/swisslandstats-geopy/commit/6ac4b289ab7242a1f1b82a61f879ccaaab3ccf9a).

Martí
