### 2019-07-14 martibosch/swisslandstats-geopy#7 @martibosch
I have added (to the README, documentation and paper) the target audience and examples of its use in the literature in [e560582](https://github.com/martibosch/swisslandstats-geopy/commit/e560582a62ddc0f5f03589d32e19c7769c8a771c).

Martí
