### 2019-07-16 #1511 @weikang9009
Hi @leouieda, the package was updated and [a new version (V0.7.1) was released on github and pypi 7 days ago by @martibosch](https://github.com/martibosch/swisslandstats-geopy/releases/tag/v0.7.1). Since this most recent version does not match the one given in the review points (shown below), I am not sure how to deal with this?

> - [ ] **Version:** Does the release version given match the GitHub release (v0.6.1)?
