### 2019-07-17 #1511 @leouieda
@darribas thanks for the update :+1:

@weikang9009 that happens when there is a release while the review is ongoing. As long as the version is close and matches the Github release/tag, then it should be OK. We'll update the version information prior to publication.
