### 2019-07-17 #1511 @darribas
On the authorship, @martibosch has clearly contributed to the package. In fact, he's the only contributor). The second author has not. Usually, this would point to lack of substantive contribution but I could see how this is also reasonable if the second author has provided guidance, technical support, etc. Happy to check that one off if @martibosch agrees.
