### 2019-07-17 martibosch/swisslandstats-geopy#9 @darribas
[JOSE review] `cythonized` geopandas version?
The [README](https://github.com/martibosch/swisslandstats-geopy/blob/master/README.md) states that a pre-requisite for running the package is the cythonised version of `geopandas`. I don't think the [`cython` branch](https://github.com/geopandas/geopandas/tree/geopandas-cython) has been merged. That can be a bit off-putting for people who need to relay on more stable versions so, unless it's very important, I'd make clear that it can work also with standard `geopandas`. The README could also detail the benefits of using the `dev` channel and how to do it, but showing that it's an option more than a requirement.

[openjournals/joss-reviews#1511](https://github.com/openjournals/joss-reviews/issues/1511)
