### 2019-07-17 martibosch/swisslandstats-geopy#10 @darribas
[JOSE review] Example notebook
The binder button nicely takes you to a demo that works. However, it's not clear where that notebook lives or, particularly, that it's not part of the package. Given it's a relatively light file, I'd encourage to include it on the `examples` folder of the package, where the user might expect it to be.

Similarly, it might help to add to that notebook that `!make download_data` will probably not work on Windows.

[openjournals/joss-reviews#1511](https://github.com/openjournals/joss-reviews/issues/1511)
