### 2019-07-17 martibosch/swisslandstats-geopy#11 @darribas
[JOSE review]
A few comments on the paper:

- [x] Statement of need: I think whether there's a subsection or not, it'd be helpful to start the paper with a (short) paragraph on why the package is required and what problem it solves.
- [x] The point above could also be included in the `README.md` file on top, just before the target audience.
- [x] To make the paper a bit more useful for the reader and prove its point more effectively, I think it'd help to provide short code snippets on the list of features listed at the end of the *Summary* section. I'm thinking, for every bullet point, to include a "demo snippet". For example, for the first one, it could read:

---
> * Read CSV files from the SFSO intoLandDataFrameobjects, which extend the conven-tional pandasDataFramewith additional attributes that store the coordinate referencesystem (CRS) and pixel resolution.

```python
import swisslandstats
ldf = swisslandstats.read_csv('data/AREA_NOAS04_17_181029.csv')
```
---

- [ ] Finally, I think the functionality in this paper is super cool and interesting and I wonder whether the package could be used in other contexts where data is gridded in a similar way (is there any? I know there's a lot of gridded datasets that are provided as text files but I'm not sure if the format is the same). It'd be great to elaborate a bit more on this, either stating other datasets for which this is relevant or, maybe more useful, discussing what are the characteristics that a dataset needs to fulfill.

[openjournals/joss-reviews#1511](https://github.com/openjournals/joss-reviews/issues/1511)
