### 2019-07-17 #1511 @darribas
OK that's my review for now @martibosch. Such a cool little package!!!
