### 2019-07-18 3592db2
[ci skip] standard geopandas as default in README (cython as option)
