+++ environment.yml
@@ -1,0 +1,11 @@
+name: swisslandstats
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - pip
+  - pip:
+    - swisslandstats-geopy==0.7.1
+  - geopandas=0.5.0
+  - osmnx=0.10
+  - pyproj=1.9.6
