### 2019-07-18 martibosch/swisslandstats-geopy#9 @martibosch
Hello,

I have updated the README in [3592db2](https://github.com/martibosch/swisslandstats-geopy/commit/3592db201e6837cfec197ce861970c3f90adc28b) so that it suggests the standard geopandas installation by default while still mentioning the cythonized version of geopandas as an option. I hope that this addresses the issue.

Martí
