### 2019-07-18 martibosch/swisslandstats-geopy#10 @martibosch
Hello,

I have moved the notebook and Makefile (with a note concerning Make and windows in the first cell of the notebook) into the `examples` folder of this repository and updated the README (and the Binder badge) in [e61704d](https://github.com/martibosch/swisslandstats-geopy/commit/e61704dccb27865e54f63dbfc96e92638061d46f). I hope that this addresses the issue.

Martí
