### 2019-07-18 8a7d0ab
[ci skip] added statement of need in paper and README
