### 2019-07-18 martibosch/swisslandstats-geopy#10 @darribas
Yep! that's excellent! Closing...
