### 2019-07-18 martibosch/swisslandstats-geopy#9 @darribas
Great! Closing...
