### 2019-07-18 a8f34fe
[ci skip] added demo snippets for each feature in the paper
