+++ paper/paper.md
@@ -38,9 +38,43 @@ E       | N       | FJ85 | ... | FJ18 | AS85_17 | ... | AS18_17 |
 While the above structure allows storing the different survey periods into a single file which can be straightforwardly read by most data analysis packages, libraries to process geographical raster data aree rarely capable of processing such format. Therefore, the aim of the proposed library, swisslandstats-geopy is to provide an extended pandas `DataFrame` interface [@mckinney2010data] to the table-like LULC inventory provided by the SFSO, which includes the following features:
 
 * Read CSV files from the SFSO into `LandDataFrame` objects, which extend the conventional pandas `DataFrame` with additional attributes that store the coordinate reference system (CRS) and pixel resolution.
+
+    ```python
+    import swisslandstats
+    ldf = swisslandstats.read_csv('data/AREA_NOAS04_17_181029.csv')
+    ```
+
 * Export the categorical LULC columns into NumPy arrays [@van2011numpy] and GeoTIFF files
+
+    ```python
+    ldf.to_ndarray('AS18_4')
+    ```
+    
+    ```
+    array([[3, 0, 0, ..., 0, 0, 0],
+          [0, 0, 0, ..., 0, 0, 0],
+          [0, 0, 0, ..., 0, 0, 0],
+          ...,
+          [3, 3, 2, ..., 0, 0, 0],
+          [3, 3, 3, ..., 0, 0, 0],
+          [3, 2, 2, ..., 0, 0, 0]], dtype=uint8)
+    ```
+
+
 * Clip `LandDataFrame` instances by Shapely vector geometries [@gillies2007shapely] or directly by region names. The latter uses OSMnx [@boeing2017osmnx] to retrieve the geometries from the [OpenStreetMap](https://www.openstreetmap.org/)
+
+    ```python
+    lausanne_ldf = ldf.clip_by_nominatim('Lausanne, Vaud')
+    ```
+    
 * Plot the categorical LULC columns as raster images
+
+    ```python
+    ax = lausanne_ldf.plot('AS09R_4')
+    ax.tick_params(axis='x', rotation=45)
+    ``` 
+    
+    ![Resulting plot.](landstats_lausanne.png)
 
 The target audience of swisslandstats-geopy is researchers and developers in environmental sciences and GIS, who intend to produce repeatable and reproducible computational workflows that make use of the LULC inventory provided by the SFSO. Examples of applications of the library in the academic literature include the assessment of the carbon sequestration for the canton of Vaud [@jaligot2019assessing, see also [a dedicated GitHub repository](https://github.com/martibosch/carbon-sequestration-vaud) with the materials necessary to reproduce the results], and the evaluation of the spatio-temporal patterns of LULC change in the urban agglomerations of Zurich, Bern and Lausanne [@bosch2019spatiotemporal, see also [a dedicated GitHub repository](https://github.com/martibosch/swiss-urbanization) with the materials necessary to reproduce the results].
 
