### 2019-07-18 #1511 @martibosch
Regarding the comment on authorship: this library has been created in the context of my PhD thesis, which is why I included my thesis supervisor as second author. Nevertheless, since this paper is not central to my thesis it is not necessary that he appears here as I am indeed the sole contributor to the library. I have therefore changed the paper accordingly in [1c8f888](https://github.com/martibosch/swisslandstats-geopy/commit/1c8f888bc15aa278de99b2bd98c928390956957a)
