### 2019-07-18 #1511 @leouieda
@martibosch we don't require all authors to be package contributors since there are non-code contributions that merit authorship but aren't registered in the git log. I'll leave the final decision up to you.
