### 2019-07-19 #1511 @darribas
Yes, my question wasn't a demand but much more in line with @leouieda's view. It was just raising bringing it up...
