### 2019-07-19 martibosch/swisslandstats-geopy#11 @darribas
Excellent. Yes, I think a couple of lines reflecting on what it'd take to make this more general would be a great addition!
