+++ paper/paper.md
@@ -78,6 +78,8 @@ While the above structure allows storing the different survey periods into a sin
 
 The target audience of swisslandstats-geopy is researchers and developers in environmental sciences and GIS, who intend to produce repeatable and reproducible computational workflows that make use of the LULC inventory provided by the SFSO. Examples of applications of the library in the academic literature include the assessment of the carbon sequestration for the canton of Vaud [@jaligot2019assessing, see also [a dedicated GitHub repository](https://github.com/martibosch/carbon-sequestration-vaud) with the materials necessary to reproduce the results], and the evaluation of the spatio-temporal patterns of LULC change in the urban agglomerations of Zurich, Bern and Lausanne [@bosch2019spatiotemporal, see also [a dedicated GitHub repository](https://github.com/martibosch/swiss-urbanization) with the materials necessary to reproduce the results].
 
+On the other hand, the `settings` module of swisslandstats-geopy allows changing the CRS and pixel resolution so that the library might also be used with similarly table-like raster datasets. For instance, a dataset of phenology in the Eastern US and Canadian forests for the 1984-2013 period [@melaas2018landsat] has been processed seamlessly into a `LandDataFrame` instance. Neverhteless, although other table-like raster datasets such as the European land use/cover area frame statistical survey (LUCAS) [@ballin2018redesign] could be instantiated as `LandDataFrames`, the data could not be converted to NumPy arrays, likely because of the CRS and how their grid is sampled. Further exploration of the characteristics that a table-like raster dataset must fulfill in order to be seamlessly procesed within swisslandstats-geopy could signficantly enhance the reusability of the library.
+
 
 # Availability
 
