+++ paper/paper.bib
@@ -45,6 +45,14 @@
   howpublished={Available from (in French) \url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/sol-utilisation-couverture/statistique-suisse-superficie/nomenclature-standard.assetdetail.4103540.html}. Accessed: 18 April 2019},
 }
 
+@misc{ballin2018redesign,
+  title={Redesign sample for Land Use/Cover Area frame Survey (LUCAS) 2018},
+  author={Ballin, Marco and Barcaroli, Giulio and Masselli, Mauro and Scarn{\'\o}, Marco},
+  year={2018},
+  publisher={Eurostat: statistical working papers},
+  doi={10.2785/132365},
+}
+
 @inproceedings{jupyter2018binder,
   title={Binder 2.0 - Reproducible, interactive, sharable environments for science at scale},
   author={{M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing},
@@ -52,6 +60,14 @@
   pages={113-120},
   year={2018},
   doi={10.25080/Majora-4af1f417-011}
+}
+
+@misc{melaas2018landsat,
+  title={Landsat-derived Spring and Autumn Phenology, Eastern US - Canadian Forests, 1984-2013},
+  author={Melaas, Eli K. and Friedl, Mark A. and Sulla-Menashe, Damien},
+  year={2018},
+  publisher={ORNL Distributed Active Archive Center},
+  doi={10.3334/ORNLDAAC/1570},  
 }
 
 @article{jaligot2019assessing,
