### 2019-07-20 martibosch/swisslandstats-geopy#11 @martibosch
I have included a dedicated paragraph with references to the datasets in [d1a9913](https://github.com/martibosch/swisslandstats-geopy/commit/d1a991352bd46054c8c6aea43eb0566776ca441a). It is 11 lines (a bit more than "a couple"), but I could make it shorter if needed.
