### 2019-07-31 #1511 @martibosch
well I believe that in this case, leaving it to one author is more coherent with the [JOSS autorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship).
