### 2019-08-01 #1511 @leouieda
:wave: @weikang9009 and @darribas please have a final look to see if all your comments have been addressed.

@darribas I see that there are some unchecked items in your reviewer checklist. Are these still in progress or have they been satisfied with the recent changes?
