### 2019-08-12 075df26
Minor fixes to the JOSS paper
Corrected a typo and other misc fixes to the text.

#openjournals/joss-reviews#1511
