### 2019-08-12 #1511 @leouieda
@martibosch I submitted a PR with a few minor fixes to the JOSS paper. Please let me know when you merge it and we can move on with the publication.
