### 2019-08-13 #1511 @martibosch
@leouieda thank you for your corrections. I have already reviewed and merged the PR with your fixes.
Martí
