### 2019-08-30 #1511 @labarba
Is this submission ready for the final checks before acceptance?
