### 2019-08-31 b08173d
Fix caption and make reference to the figure 1
The figure won't show up in that exact position so the caption "resulting plot"
is not particularly informative. I added a reference to the figure in the text
and a more explanatory caption.

openjournals/joss-reviews#1511
