### 2019-08-31 #1511 @leouieda
@labarba thank you for the nudge, it's completely my fault that this hasn't moved along. I made one final editorial PR to the paper. After that, this should be good to go.
