### 2019-09-04 #1511 @leouieda
@martibosch  I'm happy say that your paper is ready for publication in JOSS! :tada: Here are the final steps you need to take:

- Make a release with the latest changes from the review and post the version number here.
- Archive the release on Zenodo/figshare and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).
