### 2019-09-04 #1511 @martibosch
> - [x] Double check authors and affiliations (including ORCIDs)

Done

> - [ ] Make a release with the latest changes from the review and post the version number here.

The latest release is 0.7.1, and all the subsequent commits have been amendments to the paper/README and including the "examples" folder in the GitHub repository (however, such folder is not included in the MANIFEST), therefore nothing would change in an hypothetical new release (well, the README would be different - but that does not seem to justify a release in semantic versioning). Shall we leave it like that, or release a patch (i.e., 0.7.2) with the above changes?

> - [ ] Archive the release on Zenodo/figshare and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).

Once the second task is done, I will proceed with the third and last one.
