### 2019-09-05 #1511 @leouieda
@martibosch it would be best to have a patch release if possible. You might want to take this opportunity to set the citation information for the package to the JOSS paper (the doi will be 10.21105.joss.01511) and have it registered in the release.
