### 2019-09-05 6632d2c
[ci skip] added JOSS badge and citation info to README
