+++ swisslandstats/__init__.py
@@ -2,4 +2,4 @@ from .dataframe import *
 from .geometry import *
 from .plotting import *
 
-__version__ = '0.7.1'
+__version__ = '0.7.2'
