### 2019-09-05 #1511 @martibosch
I have released 0.7.2 and archived it to Zenodo https://doi.org/10.5281/zenodo.3387312 (with the matching title and author list).
