### 2019-09-11 #1511 @Kevin-Mattheus-Moerman
@martibosch can you check/work on the following points in relation to your paper:

- [ ] We are about to process this submission for acceptance. Can you please thoroughly read and double check the paper? Also ensure all author names and affiliations are accurate and that acknowledgements to contributors and funding sources have been added.
- [ ] Can you remove the acronyms from your affiliations (I would keep them if they are used in the text but since they are not you can remove them).
- [ ] In "..are rarely capable of processing such format.". Please consider using "formats" or "a format".
- [ ] Does NumPy need its own independent reference?
- [ ] Although usable are you able to improve the resolution of figure 1 (its a bit fuzzy)?
- [ ] Fix the word nevertheless in"Neverhteless, although other...."
- [ ] Once you have complete the above, please run `@whedon generate pdf` here to update your proofs
