### 2019-09-11 #1511 @Kevin-Mattheus-Moerman
@leouieda can you verify the above points and ping me once we are ready to proceed? Thanks
