### 2019-09-11 996ddc0
[ci skip] final corrections to the paper
