### 2019-09-11 #1511 @martibosch
* [x]  We are about to process this submission for acceptance. Can you please thoroughly read and double check the paper? Also ensure all author names and affiliations are accurate and that acknowledgements to contributors and funding sources have been added.

* [x]  Can you remove the acronyms from your affiliations (I would keep them if they are used in the text but since they are not you can remove them).

* [x]  In "..are rarely capable of processing such format.". Please consider using "formats" or "a format".
* [x]  Does NumPy need its own independent reference?

Well, I believe that since pandas, shapely and OSMnx have their own independent reference, NumPy should be treated equally. But this can be changed if you consider it more appropriate.

* [x]  Although usable are you able to improve the resolution of figure 1 (its a bit fuzzy)?

I have changed the respective code snippets in the paper so that the full extent of the canton of Vaud is plotted. I believe it might be a bit too big now. I can amend the commit with a smaller figure if you consider it more appropriate.

* [x]  Fix the word nevertheless in"Neverhteless, although other...."
* [x]  Once you have complete the above, please run `@whedon generate pdf` here to update your proofs
