### 2019-09-11 #1511 @leouieda
> Well, I believe that since pandas, shapely and OSMnx have their own independent reference, NumPy should be treated equally. But this can be changed if you consider it more appropriate.

I agree that numpy deserves a citation.

> I have changed the respective code snippets in the paper so that the full extent of the canton of Vaud is plotted. I believe it might be a bit too big now. I can amend the commit with a smaller figure if you consider it more appropriate.

The figure size is fine but the resolution is a bit low (which I think was the original criticism). If you could please save it with a higher DPI, that would solve the problem.
