### 2019-09-13 #1511 @leouieda
Thank you @martibosch, the figure looks much better now :+1:

@Kevin-Mattheus-Moerman I believe all points have been addressed and we should be ready to move forward.
