### 2019-09-15 #1511 @arfon
> @arfon, still a 404 it seems, on the DOI. I'll check again later.

Looks like there was a minor issue with the paper (no affiliation index) which for some reason Whedon failed silently on. I've fixed this now.
